Move registry out of graphql module
