+++ proxy/src/registry.rs
